[PRODCRE-1746] Make keys folder self-contained. (#21130)
* Minor.

* Minor.

* Minor.

* Minor.

* Minor.

* Minor.

* Minor.
