+++ core/store/models/common.go
@@ -3,7 +3,6 @@ package models
 import (
 	"bytes"
 	"database/sql/driver"
-	"encoding/hex"
 	"encoding/json"
 	stderrors "errors"
 	"fmt"
@@ -270,83 +269,6 @@ func Merge(inputs ...JSON) (JSON, error) {
 	}
 
 	return JSON{Result: gjson.ParseBytes(bytes)}, nil
-}
-
-// Explicit type indicating a 32-byte sha256 hash
-type Sha256Hash [32]byte
-
-var EmptySha256Hash = new(Sha256Hash)
-
-// MarshalJSON converts a Sha256Hash to a JSON byte slice.
-func (s Sha256Hash) MarshalJSON() ([]byte, error) {
-	return json.Marshal(s.String())
-}
-
-// UnmarshalJSON converts a bytes slice of JSON to a TaskType.
-func (s *Sha256Hash) UnmarshalJSON(input []byte) error {
-	var shaHash string
-	if err := json.Unmarshal(input, &shaHash); err != nil {
-		return err
-	}
-
-	sha, err := Sha256HashFromHex(shaHash)
-	if err != nil {
-		return err
-	}
-
-	*s = sha
-	return nil
-}
-
-func Sha256HashFromHex(x string) (Sha256Hash, error) {
-	bs, err := hex.DecodeString(x)
-	if err != nil {
-		return Sha256Hash{}, err
-	}
-	var hash Sha256Hash
-	copy(hash[:], bs)
-	return hash, nil
-}
-
-func MustSha256HashFromHex(x string) Sha256Hash {
-	bs, err := hex.DecodeString(x)
-	if err != nil {
-		panic(err)
-	}
-	var hash Sha256Hash
-	copy(hash[:], bs)
-	return hash
-}
-
-func (s Sha256Hash) String() string {
-	return hex.EncodeToString(s[:])
-}
-
-func (s *Sha256Hash) MarshalText() ([]byte, error) {
-	return []byte(s.String()), nil
-}
-
-func (s *Sha256Hash) UnmarshalText(bs []byte) (err error) {
-	*s, err = Sha256HashFromHex(string(bs))
-	return
-}
-
-func (s *Sha256Hash) Scan(value any) error {
-	bytes, ok := value.([]byte)
-	if !ok {
-		return errors.Errorf("Failed to unmarshal Sha256Hash value: %v", value)
-	}
-	if s == nil {
-		*s = Sha256Hash{}
-	}
-	copy((*s)[:], bytes)
-	return nil
-}
-
-func (s Sha256Hash) Value() (driver.Value, error) {
-	b := make([]byte, 32)
-	copy(b, s[:])
-	return b, nil
 }
 
 // ServiceHeader is an HTTP header to include in POST to log service.
